+++ frontend/src/app/App.tsx
@@ -1,7 +1,8 @@
 import { Admin } from "react-admin";
 
 import { Dashboard } from "../features/dashboard/Dashboard";
+import { authProvider } from "../services/authProvider";
 
 export function App() {
-  return <Admin dashboard={Dashboard} disableTelemetry />;
+  return <Admin authProvider={authProvider} dashboard={Dashboard} disableTelemetry />;
 }